requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
jinja2==3.1.4
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1
numpy==1.26.4
Pillow==10.3.0
pytest==8.2.0
httpx==0.27.0
